Bug 852583 - add an additional debug target: bluetooth

+++ android/main-emulator.c
@@ -30,7 +30,7 @@
 #include <android/avd/util.h>
 
 /* Required by android/utils/debug.h */
-int android_verbose;
+unsigned long long android_verbose;
 
 
 #define DEBUG 1
